Begin splitting up data-utils module
