+++ src/components/data/DataTable.js
@@ -3,7 +3,8 @@ import { Fragment, useEffect, useRef, useState } from 'react'
 import { b, div, h, span } from 'react-hyperscript-helpers'
 import { AutoSizer } from 'react-virtualized'
 import { ButtonPrimary, ButtonSecondary, Checkbox, Clickable, ClipboardButton, DeleteConfirmationModal, fixedSpinnerOverlay, Link, RadioButton } from 'src/components/common'
-import { concatenateAttributeNames, EditDataLink, entityAttributeText, EntityRenamer, HeaderOptions, renderDataCell, SingleEntityEditor } from 'src/components/data/data-utils'
+import { concatenateAttributeNames } from 'src/components/data/attribute-utils'
+import { EditDataLink, entityAttributeText, EntityRenamer, HeaderOptions, renderDataCell, SingleEntityEditor } from 'src/components/data/data-utils'
 import RenameColumnModal from 'src/components/data/RenameColumnModal'
 import { allSavedColumnSettingsEntityTypeKey, allSavedColumnSettingsInWorkspace, ColumnSettingsWithSavedColumnSettings, decodeColumnSettings } from 'src/components/data/SavedColumnSettings'
 import { icon } from 'src/components/icons'
