+++ src/components/data/_tests/data-utils.test.js
@@ -1,83 +1,10 @@
 import '@testing-library/jest-dom'
 
-import { fireEvent, render } from '@testing-library/react'
+import { render } from '@testing-library/react'
 import _ from 'lodash/fp'
-import { h } from 'react-hyperscript-helpers'
-import { AttributeInput, AttributeTypeInput, concatenateAttributeNames, convertAttributeValue, entityAttributeText, getAttributeType, prepareAttributeForUpload, renderDataCell } from 'src/components/data/data-utils'
+import { entityAttributeText, prepareAttributeForUpload, renderDataCell } from 'src/components/data/data-utils'
 import * as Utils from 'src/libs/utils'
 
-
-describe('concatenateAttributeNames', () => {
-  it('works with two empty arrays', () => {
-    const attrList1 = []
-    const attrList2 = []
-    const expected = []
-    expect(concatenateAttributeNames(attrList1, attrList2)).toEqual(expected)
-  })
-  it('works with left empty array', () => {
-    const attrList1 = []
-    const attrList2 = ['aaa', 'bbb']
-    const expected = ['aaa', 'bbb']
-    expect(concatenateAttributeNames(attrList1, attrList2)).toEqual(expected)
-  })
-  it('works with right empty array', () => {
-    const attrList1 = ['aaa', 'bbb']
-    const attrList2 = []
-    const expected = ['aaa', 'bbb']
-    expect(concatenateAttributeNames(attrList1, attrList2)).toEqual(expected)
-  })
-  it('returns only unique attribute names', () => {
-    const attrList1 = ['namespace:aaa', 'namespace:bbb']
-    const attrList2 = ['namespace:bbb', 'namespace:ccc']
-    const expected = ['namespace:aaa', 'namespace:bbb', 'namespace:ccc']
-    expect(concatenateAttributeNames(attrList1, attrList2)).toEqual(expected)
-  })
-  it('works when arrays are equal', () => {
-    const attrList1 = ['aaa', 'bbb', 'ccc', 'ddd']
-    const attrList2 = ['aaa', 'bbb', 'ccc', 'ddd']
-    const expected = ['aaa', 'bbb', 'ccc', 'ddd']
-    expect(concatenateAttributeNames(attrList1, attrList2)).toEqual(expected)
-  })
-  it('sorts attribute names', () => {
-    const attrList1 = ['namespace:ccc', 'namespace:aaa', 'ddd']
-    const attrList2 = ['namespace:bbb', 'aaa']
-    const expected = ['aaa', 'ddd', 'namespace:aaa', 'namespace:bbb', 'namespace:ccc']
-    expect(concatenateAttributeNames(attrList1, attrList2)).toEqual(expected)
-  })
-})
-
-describe('getAttributeType', () => {
-  it('returns type of attribute value', () => {
-    expect(getAttributeType('value')).toEqual({ type: 'string', isList: false })
-    expect(getAttributeType(3)).toEqual({ type: 'number', isList: false })
-    expect(getAttributeType(false)).toEqual({ type: 'boolean', isList: false })
-    expect(getAttributeType({
-      entityType: 'thing',
-      entityName: 'thing_one'
-    })).toEqual({ type: 'reference', isList: false })
-
-    expect(getAttributeType({ items: ['a', 'b', 'c'], itemsType: 'AttributeValue' })).toEqual({ type: 'string', isList: true })
-    expect(getAttributeType({ items: [1, 2, 3], itemsType: 'AttributeValue' })).toEqual({ type: 'number', isList: true })
-    expect(getAttributeType({ items: [true, false], itemsType: 'AttributeValue' })).toEqual({ type: 'boolean', isList: true })
-    expect(getAttributeType({
-      items: [
-        { entityType: 'thing', entityName: 'thing_one' },
-        { entityType: 'thing', entityName: 'thing_two' }
-      ],
-      itemsType: 'EntityReference'
-    })).toEqual({ type: 'reference', isList: true })
-
-    expect(getAttributeType({ key: 'value' })).toEqual({ type: 'json', isList: false })
-    expect(getAttributeType(['a', 'b', 'c'])).toEqual({ type: 'json', isList: false })
-    expect(getAttributeType([{ idx: 0 }, { idx: 1 }, { idx: 2 }])).toEqual({ type: 'json', isList: false })
-  })
-
-  it('returns string for null values', () => {
-    expect(getAttributeType(null)).toEqual({ type: 'string', isList: false })
-    expect(getAttributeType({ items: [], itemsType: 'AttributeValue' })).toEqual({ type: 'string', isList: true })
-    expect(getAttributeType({ items: [null], itemsType: 'AttributeValue' })).toEqual({ type: 'string', isList: true })
-  })
-})
 
 describe('entityAttributeText', () => {
   describe('basic data types', () => {
@@ -264,94 +191,6 @@ describe('renderDataCell', () => {
   })
 })
 
-describe('convertAttributeValue', () => {
-  it('converts between different attribute types', () => {
-    expect(convertAttributeValue('42', 'number')).toEqual(42)
-    expect(convertAttributeValue('a_string', 'number')).toEqual(0)
-    expect(convertAttributeValue('a_string', 'boolean')).toEqual(true)
-    expect(convertAttributeValue('a_string', 'reference', 'thing')).toEqual({ entityType: 'thing', entityName: 'a_string' })
-    expect(convertAttributeValue('a_string', 'json')).toEqual({ value: 'a_string' })
-
-    expect(convertAttributeValue(7, 'string')).toEqual('7')
-    expect(convertAttributeValue(7, 'boolean')).toEqual(true)
-    expect(convertAttributeValue(7, 'reference', 'thing')).toEqual({ entityType: 'thing', entityName: '7' })
-    expect(convertAttributeValue(7, 'json')).toEqual({ value: 7 })
-
-    expect(convertAttributeValue(true, 'string')).toEqual('true')
-    expect(convertAttributeValue(true, 'number')).toEqual(1)
-    expect(convertAttributeValue(false, 'reference', 'thing')).toEqual({ entityType: 'thing', entityName: 'false' })
-    expect(convertAttributeValue(true, 'json')).toEqual({ value: true })
-
-    expect(convertAttributeValue({ entityType: 'thing', entityName: 'thing_one' }, 'string')).toEqual('thing_one')
-    expect(convertAttributeValue({ entityType: 'thing', entityName: 'thing_one' }, 'number')).toEqual(0)
-    expect(convertAttributeValue({ entityType: 'thing', entityName: 'thing_one' }, 'boolean')).toEqual(true)
-
-    expect(convertAttributeValue({ key: 'value' }, 'string')).toEqual('')
-    expect(convertAttributeValue({ key: 'value' }, 'number')).toEqual(0)
-    expect(convertAttributeValue({ key: 'value' }, 'boolean')).toEqual(true)
-    expect(convertAttributeValue({ key: 'value' }, 'reference', 'thing')).toEqual({ entityType: 'thing', entityName: '' })
-    expect(convertAttributeValue({ key: 'value' }, 'json')).toEqual({ key: 'value' })
-
-    expect(() => convertAttributeValue('abc', 'notatype')).toThrow('Invalid attribute type "notatype"')
-  })
-
-  it('throws an error if attempting to convert to a reference without an entity type', () => {
-    expect(() => convertAttributeValue('thing_one', 'reference')).toThrowError()
-  })
-
-  it('converts each value of lists', () => {
-    expect(convertAttributeValue({ items: ['42', 'value'], itemsType: 'AttributeValue' }, 'number')).toEqual({ items: [42, 0], itemsType: 'AttributeValue' })
-
-    expect(convertAttributeValue({
-      items: [
-        { entityType: 'thing', entityName: 'thing_one' },
-        { entityType: 'thing', entityName: 'thing_two' }
-      ],
-      itemsType: 'EntityReference'
-    }, 'string')).toEqual({ items: ['thing_one', 'thing_two'], itemsType: 'AttributeValue' })
-  })
-
-  it('adds itemsType to reference lists', () => {
-    expect(convertAttributeValue({ items: ['thing_one', 'thing_two'], itemsType: 'AttributeValue' }, 'reference', 'thing')).toEqual({
-      items: [
-        { entityType: 'thing', entityName: 'thing_one' },
-        { entityType: 'thing', entityName: 'thing_two' }
-      ],
-      itemsType: 'EntityReference'
-    })
-  })
-
-  it('converts lists to arrays', () => {
-    expect(convertAttributeValue({
-      items: ['a', 'b', 'c'],
-      itemsType: 'AttributeValue'
-    }, 'json')).toEqual(['a', 'b', 'c'])
-    expect(convertAttributeValue({
-      items: [1, 2, 3],
-      itemsType: 'AttributeValue'
-    }, 'json')).toEqual([1, 2, 3])
-    expect(convertAttributeValue({
-      items: [true, false, true],
-      itemsType: 'AttributeValue'
-    }, 'json')).toEqual([true, false, true])
-  })
-
-  it('converts arrays to lists', () => {
-    expect(convertAttributeValue(['a', 'b', 'c'], 'string')).toEqual({
-      items: ['a', 'b', 'c'],
-      itemsType: 'AttributeValue'
-    })
-    expect(convertAttributeValue(['1', '2', '3'], 'number')).toEqual({
-      items: [1, 2, 3],
-      itemsType: 'AttributeValue'
-    })
-    expect(convertAttributeValue(['true', 'false', 'true'], 'boolean')).toEqual({
-      items: [true, false, true],
-      itemsType: 'AttributeValue'
-    })
-  })
-})
-
 describe('prepareAttributeForUpload', () => {
   it('trims string values', () => {
     expect(prepareAttributeForUpload('foo ')).toEqual('foo')
@@ -404,307 +243,6 @@ describe('prepareAttributeForUpload', () => {
     })).toEqual({
       items: [1, 2, 3],
       itemsType: 'AttributeValue'
-    })
-  })
-})
-
-describe('AttributeTypeInput', () => {
-  it('renders radio buttons for available types', () => {
-    const { getAllByRole } = render(h(AttributeTypeInput, {
-      value: { type: 'string' },
-      onChange: jest.fn()
-    }))
-
-    const radioButtons = getAllByRole('radio')
-    const labels = Array.from(radioButtons).map(el => el.labels[0].textContent)
-
-    expect(radioButtons.length).toBe(4)
-    expect(labels).toEqual(['String', 'Reference', 'Number', 'Boolean'])
-  })
-
-  it('renders a radio button for JSON type if requested', () => {
-    const { queryByLabelText } = render(h(AttributeTypeInput, {
-      value: { type: 'string' },
-      onChange: jest.fn(),
-      showJsonTypeOption: true
-    }))
-
-    const jsonRadioButton = queryByLabelText('JSON')
-    expect(jsonRadioButton).not.toBeNull()
-  })
-
-  it('calls onChange callback when a type is selected', () => {
-    const onChange = jest.fn()
-    const { getByLabelText } = render(h(AttributeTypeInput, {
-      value: { type: 'string' },
-      onChange
-    }))
-
-    fireEvent.click(getByLabelText('Number'))
-
-    expect(onChange).toHaveBeenCalledWith({ type: 'number' })
-  })
-
-  describe('references', () => {
-    it('renders an entity types menu for reference values', () => {
-      const { getByLabelText } = render(h(AttributeTypeInput, {
-        value: { type: 'reference', entityType: 'foo' },
-        entityTypes: ['foo', 'bar', 'baz'],
-        onChange: jest.fn()
-      }))
-
-      const entityTypeInput = getByLabelText('Referenced entity type:')
-      expect(entityTypeInput.getAttribute('role')).toBe('combobox')
-    })
-
-    it('selecting reference type uses the default reference entity type if one is provided', () => {
-      const onChange = jest.fn()
-      const { getByLabelText } = render(h(AttributeTypeInput, {
-        value: { type: 'string' },
-        entityTypes: ['foo', 'bar', 'baz'],
-        defaultReferenceEntityType: 'baz',
-        onChange
-      }))
-
-      fireEvent.click(getByLabelText('Reference'))
-
-      expect(onChange).toHaveBeenCalledWith({ type: 'reference', entityType: 'baz' })
-    })
-
-    it('selecting reference type uses the alphabetically first entity type if no default reference entity type is provided', () => {
-      const onChange = jest.fn()
-      const { getByLabelText } = render(h(AttributeTypeInput, {
-        value: { type: 'string' },
-        entityTypes: ['foo', 'bar', 'baz'],
-        onChange
-      }))
-
-      fireEvent.click(getByLabelText('Reference'))
-
-      expect(onChange).toHaveBeenCalledWith({ type: 'reference', entityType: 'bar' })
-    })
-  })
-})
-
-describe('AttributeInput', () => {
-  describe('type buttons', () => {
-    it('renders radio buttons for available types', () => {
-      const { getAllByRole } = render(h(AttributeInput, {
-        onChange: jest.fn()
-      }))
-
-      const radioButtons = getAllByRole('radio')
-      const labels = Array.from(radioButtons).map(el => el.labels[0].textContent)
-
-      expect(radioButtons.length).toBe(4)
-      expect(labels).toEqual(['String', 'Reference', 'Number', 'Boolean'])
-    })
-
-    it('renders an entity types menu for reference values', () => {
-      const { getByLabelText } = render(h(AttributeInput, {
-        value: { entityType: 'foo', entityName: 'foo_1' },
-        entityTypes: ['foo', 'bar', 'baz'],
-        onChange: jest.fn()
-      }))
-
-      const entityTypeInput = getByLabelText('Referenced entity type:')
-      expect(entityTypeInput.getAttribute('role')).toBe('combobox')
-    })
-
-    it('renders a radio button for JSON type for JSON values', () => {
-      const { queryByLabelText } = render(h(AttributeInput, {
-        value: { key: 'value' },
-        onChange: jest.fn()
-      }))
-
-      const jsonRadioButton = queryByLabelText('JSON')
-      expect(jsonRadioButton).not.toBeNull()
-    })
-
-    it('renders a radio button for JSON type if requested', () => {
-      const { queryByLabelText } = render(h(AttributeInput, {
-        onChange: jest.fn(),
-        showJsonTypeOption: true
-      }))
-
-      const jsonRadioButton = queryByLabelText('JSON')
-      expect(jsonRadioButton).not.toBeNull()
-    })
-  })
-
-  describe('selecting a type', () => {
-    it('converts value to the selected type', () => {
-      const onChange = jest.fn()
-      const { getByLabelText } = render(h(AttributeInput, {
-        value: '123',
-        onChange
-      }))
-
-      fireEvent.click(getByLabelText('Number'))
-
-      expect(onChange).toHaveBeenCalledWith(123)
-    })
-
-    it('if an initial value is provided, converts initial value to the selected type', () => {
-      // When editing an attribute, this allows previewing the effect of a type change without
-      // performing the lossy conversion.
-      const onChange = jest.fn()
-      const { getByLabelText } = render(h(AttributeInput, {
-        value: '123',
-        initialValue: '456',
-        onChange
-      }))
-
-      fireEvent.click(getByLabelText('Number'))
-
-      expect(onChange).toHaveBeenCalledWith(456)
-    })
-  })
-
-  describe('renders a value input based on the attribute type', () => {
-    it('renders a text input for string attributes', () => {
-      const onChange = jest.fn()
-      const { getByLabelText } = render(h(AttributeInput, {
-        value: 'value',
-        onChange
-      }))
-      const valueInput = getByLabelText('New value')
-
-      expect(valueInput.tagName).toBe('INPUT')
-      expect(valueInput.type).toBe('text')
-      expect(valueInput.value).toBe('value')
-
-      fireEvent.change(valueInput, { target: { value: 'newvalue' } })
-      expect(onChange).toHaveBeenCalledWith('newvalue')
-    })
-
-    it('renders a text input for reference attributes', () => {
-      const onChange = jest.fn()
-      const { getByLabelText } = render(h(AttributeInput, {
-        value: { entityType: 'thing', entityName: 'thing_one' },
-        onChange
-      }))
-      const valueInput = getByLabelText('New value')
-
-      expect(valueInput.tagName).toBe('INPUT')
-      expect(valueInput.type).toBe('text')
-      expect(valueInput.value).toBe('thing_one')
-
-      fireEvent.change(valueInput, { target: { value: 'thing_two' } })
-      expect(onChange).toHaveBeenCalledWith({ entityType: 'thing', entityName: 'thing_two' })
-    })
-
-    it('renders a number input for number attributes', () => {
-      const onChange = jest.fn()
-      const { getByLabelText } = render(h(AttributeInput, {
-        value: 123,
-        onChange
-      }))
-      const valueInput = getByLabelText('New value')
-
-      expect(valueInput.tagName).toBe('INPUT')
-      expect(valueInput.type).toBe('number')
-      expect(valueInput.value).toBe('123')
-
-      fireEvent.change(valueInput, { target: { value: '456' } })
-      expect(onChange).toHaveBeenCalledWith(456)
-    })
-
-    it('renders a switch for boolean attributes', () => {
-      const onChange = jest.fn()
-      const { getByLabelText } = render(h(AttributeInput, {
-        value: true,
-        onChange
-      }))
-      const valueInput = getByLabelText('New value')
-
-      expect(valueInput.tagName).toBe('INPUT')
-      expect(valueInput.type).toBe('checkbox')
-      expect(valueInput.checked).toBe(true)
-
-      fireEvent.click(valueInput)
-      expect(onChange).toHaveBeenCalledWith(false)
-    })
-  })
-
-  describe('lists', () => {
-    it('renders a checkbox for list status', () => {
-      const { getByLabelText } = render(h(AttributeInput, {
-        value: { itemsType: 'AttributeValue', items: ['foo', 'bar', 'baz'] },
-        onChange: jest.fn()
-      }))
-
-      const listCheckbox = getByLabelText('Value is a list')
-      expect(listCheckbox.getAttribute('role')).toBe('checkbox')
-      expect(listCheckbox.getAttribute('aria-checked')).toBe('true')
-    })
-
-    it('it converts single value to a list when list checkbox is checked', () => {
-      const onChange = jest.fn()
-      const { getByLabelText } = render(h(AttributeInput, {
-        value: 'foo',
-        onChange
-      }))
-
-      const listCheckbox = getByLabelText('Value is a list')
-      fireEvent.click(listCheckbox)
-
-      expect(onChange).toHaveBeenCalledWith({ itemsType: 'AttributeValue', items: ['foo'] })
-    })
-
-    it('it converts list to a single value when list checkbox is unchecked', () => {
-      const onChange = jest.fn()
-      const { getByLabelText } = render(h(AttributeInput, {
-        value: { itemsType: 'AttributeValue', items: ['foo', 'bar', 'baz'] },
-        onChange
-      }))
-
-      const listCheckbox = getByLabelText('Value is a list')
-      fireEvent.click(listCheckbox)
-
-      expect(onChange).toHaveBeenCalledWith('foo')
-    })
-
-    it('renders an input for each list item', () => {
-      const onChange = jest.fn()
-      const { getAllByLabelText } = render(h(AttributeInput, {
-        value: { itemsType: 'AttributeValue', items: ['foo', 'bar', 'baz'] },
-        onChange
-      }))
-
-      const valueInputs = getAllByLabelText(/^List value \d+/)
-      expect(valueInputs.length).toBe(3)
-      expect(Array.from(valueInputs).map(el => el.value)).toEqual(['foo', 'bar', 'baz'])
-
-      fireEvent.change(valueInputs[1], { target: { value: 'qux' } })
-      expect(onChange).toHaveBeenCalledWith({ itemsType: 'AttributeValue', items: ['foo', 'qux', 'baz'] })
-    })
-
-    it('renders buttons to remove items from list', () => {
-      const onChange = jest.fn()
-      const { getAllByLabelText } = render(h(AttributeInput, {
-        value: { itemsType: 'AttributeValue', items: ['foo', 'bar', 'baz'] },
-        onChange
-      }))
-
-      const removeButtons = getAllByLabelText(/^Remove list value \d+/)
-      expect(removeButtons.length).toBe(3)
-
-      fireEvent.click(removeButtons[1])
-      expect(onChange).toHaveBeenCalledWith({ itemsType: 'AttributeValue', items: ['foo', 'baz'] })
-    })
-
-    it('renders button to add item to list', () => {
-      const onChange = jest.fn()
-      const { getByText } = render(h(AttributeInput, {
-        value: { itemsType: 'AttributeValue', items: ['foo', 'bar', 'baz'] },
-        onChange
-      }))
-
-      const addButton = getByText('Add item')
-      fireEvent.click(addButton)
-
-      expect(onChange).toHaveBeenCalledWith({ itemsType: 'AttributeValue', items: ['foo', 'bar', 'baz', ''] })
     })
   })
 })
